third_party/blink/web_tests/platform/generic/media/W3C/video/readyState/readyState_during_loadedmetadata-expected.txt

video.readyState should be >= HAVE_METADATA during loadedmetadata event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
